Remove Support button on Billing page and introduce a hook
