+++ packages/twenty-front/src/modules/support/components/SupportDropdown.tsx
@@ -26,7 +26,7 @@ export const SupportDropdown = () => {
       dropdownId={dropdownId}
       dropdownPlacement="top-start"
       dropdownOffset={{ x: 0, y: -28 }}
-      clickableComponent={<SupportButton isClickable={false} />}
+      clickableComponent={<SupportButton />}
       dropdownComponents={
         <DropdownMenu width="160px">
           <DropdownMenuItemsContainer>
